Create stabilization
This will be the main code used in the arduino, some of the syntax is not yet perfected but I'd like to get the main functions in here. Please update if you see any changes that need to be made or if you are able to clear up some of the syntax.