Update and rename stabilization to Independent Stabilization
Removed all of the commented out statements and changed a few things so that the code should run independently and re-orient itself to the position calculated when the code first ran on the launch pad. Will have to test to make sure it is working properly.

The one issue on both this code and the Active roll control with stabilization code is that we need to figure out another way to determine how far away we are from our desired angular orientation. Currently if we have values close to 0 or 360 degrees we could run into some big problems. Luckily the code loops and resets the servos so quickly that it still may not make a big difference.